requirements.txt
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
configobj==5.0.6
firebase-admin==5.0.2
Flask==2.0.1
flask-ngrok==0.0.25
future==0.18.2
google-api-core==1.31.2
google-api-python-client==2.20.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.0.0
google-cloud-firestore==2.3.1
google-cloud-storage==1.42.1
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
grpcio==1.40.0
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
msgpack==1.0.2
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymessenger==0.0.7.0
pyparsing==2.4.7
pyTelegramBotAPI==4.0.0
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.6
viberbot==1.0.12
Werkzeug==2.0.1
